refactor : SelectSeats 모델로 이동 및 도메인 -> 모델로 이름 변경
