+++ app/src/main/java/woowacourse/movie/model/MovieScheduleGenerator.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.domain
+package woowacourse.movie.model
 
 import java.time.DayOfWeek
 import java.time.LocalDate
