+++ app/src/main/java/woowacourse/movie/model/TicketMaker.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.domain
+package woowacourse.movie.model
 
 import woowacourse.movie.uiModel.TicketUIModel
