+++ app/src/test/java/woowacourse/movie/MovieScheduleGeneratorTest.kt
@@ -2,7 +2,7 @@ package woowacourse.movie
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
-import woowacourse.movie.domain.MovieScheduleGenerator
+import woowacourse.movie.model.MovieScheduleGenerator
 
 class MovieScheduleGeneratorTest {
     @Test
